moved roi calc functions back to stats.py
