+++ .dvc/.gitignore
@@ -26,3 +26,5 @@
 /tmp
 /tmp
 /tmp
+/tmp
+/tmp
