+++ suite2p/detection/stats.py
@@ -1,23 +1,18 @@
-from typing import Tuple, Optional
+from typing import Tuple, Optional, NamedTuple
 from dataclasses import dataclass, field
 from warnings import warn
 
 import numpy as np
 from numpy.linalg import norm
 
-from .utils import fitMVGaus, distance_kernel, norm_by_average
+from .utils import norm_by_average
 
 
-def mean_r_squared(y: np.ndarray, x: np.ndarray, estimator=np.median) -> float:
-    return np.mean(norm(((y - estimator(y)), (x - estimator(x))), axis=0))
-
-
-def calc_radii(dy: float, dx: float, ypix: np.ndarray, xpix: np.ndarray, lam: np.ndarray) -> Tuple[float, float]:
-    return fitMVGaus(ypix / dy, xpix / dx, lam, 2).radii
-
-
-def aspect_ratio(ry: float, rx: float, offset: float = .01) -> float:
-    return 2 * ry / (ry + rx + offset)
+def distance_kernel(radius: int) -> np.ndarray:
+    """ Returns 2D array containing geometric distance from center, with radius 'radius'"""
+    d = np.arange(-radius, radius + 1)
+    dists_2d = norm(np.meshgrid(d, d), axis=0)
+    return dists_2d
 
 
 @dataclass(frozen=True)
@@ -110,3 +105,61 @@ def roi_stats(dy: int, dx: int, stats):
         stat['footprint'] = 0 if 'footprint' not in stat else stat['footprint']
 
     return np.array(stats)
+
+
+def mean_r_squared(y: np.ndarray, x: np.ndarray, estimator=np.median) -> float:
+    return np.mean(norm(((y - estimator(y)), (x - estimator(x))), axis=0))
+
+
+def calc_radii(dy: float, dx: float, ypix: np.ndarray, xpix: np.ndarray, lam: np.ndarray) -> Tuple[float, float]:
+    return fitMVGaus(ypix / dy, xpix / dx, lam, 2).radii
+
+
+def aspect_ratio(ry: float, rx: float, offset: float = .01) -> float:
+    return 2 * ry / (ry + rx + offset)
+
+
+class EllipseData(NamedTuple):
+    mu: float
+    cov: float
+    radii: Tuple[float, float]
+    ellipse: np.ndarray
+
+    @property
+    def area(self):
+        return (self.radii[0] * self.radii[1]) ** 0.5 * np.pi
+
+
+def fitMVGaus(y, x, lam, thres=2.5, npts: int = 100) -> EllipseData:
+    """ computes 2D gaussian fit to data and returns ellipse of radius thres standard deviations.
+
+    Parameters
+    ----------
+    y : float, array
+        pixel locations in y
+    x : float, array
+        pixel locations in x
+    lam : float, array
+        weights of each pixel
+    """
+
+    # normalize pixel weights
+    lam /= lam.sum()
+
+    # mean of gaussian
+    yx = np.stack((y, x))
+    mu = (lam * yx).sum(axis=1)
+    yx = (yx - mu[:, np.newaxis]) * lam ** .5
+    cov = yx @ yx.T
+
+    # radii of major and minor axes
+    radii, evec = np.linalg.eig(cov)
+    radii = thres * np.maximum(0, np.real(radii)) ** .5
+
+    # compute pts of ellipse
+    theta = np.linspace(0, 2 * np.pi, npts)
+    p = np.stack((np.cos(theta), np.sin(theta)))
+    ellipse = (p.T * radii) @ evec.T + mu
+    radii = np.sort(radii)[::-1]
+
+    return EllipseData(mu=mu, cov=cov, radii=radii, ellipse=ellipse)
